NPCバトルコース、ステージをファイル分割した
