+++ src/js/game/npc-battle/simple-npc-battle-course.js
@@ -1,0 +1,28 @@
+// @flow
+import type {NPCBattleCourse} from "./npc-battle-course";
+import {DefaultStage} from "./npc-battle-stage";
+import type {NPCBattleStage, StageLevel} from "./npc-battle-stage";
+
+/** NPCバトルコースのシンプルな実装 */
+export class SimpleNPCBattleCourse implements NPCBattleCourse {
+  _stages: NPCBattleStage[];
+
+  /**
+   * コンストラクタ
+   *
+   * @param stages コースに含まれる全ステージ
+   */
+  constructor(stages: NPCBattleStage[]) {
+    this._stages = stages;
+  }
+
+  /** @override */
+  stage(level: StageLevel): NPCBattleStage {
+    return this._stages[level - 1] ?? DefaultStage;
+  }
+
+  /** @override */
+  lastStageLevel(): StageLevel {
+    return this._stages.length;
+  }
+}
